proxy: improve libevent handler
fixes an issue with memcached infinite looping when timeouts fire,
because we'e stacked them on top of each other. think that also lead to
a really slow memory leak.

libevent's relationship with uninitialized memory is a little annoying:
there doesn't seem to be a function to simply "initialize" it, so we
need to always check if it's initialized before checking if events are
pending.. and you shouldn't/can't just call event_del() if there isn't
an event pending, but you need to call event_del() if there is an event
pending.

this isn't quite right (Think it should be matching against 'flags'
instead of EV_READ|EV_WRITE) but I'll come back to it and move it to a
function.

should probably also rewire so backend connecting stage gets an event
with its own handler (which dequeues events/etc if they exist), which
would mean the event is always initialized so the hot path can lose
some calls at least.