proxy: API for loadable hash selectors.
FWIW: can't pass an option to the selector yet but most of the code is there
already.

This allows passing a table, which provides a "new" function, to run
initialization work for managing a custom hash selector (ie; ketama).

The new function is passed a copy of the pool table which it can use to
create any necessary data structures. The selector function must return
an index into the original pool table. See example elsewhere.