proxy: fix memory leak in ketama example
I was going to add a metatable + __gc routine for ketama, but all of the
memory is actually encapsulated in the userdata already. That gets freed
properly. What we had was an existing memory leak because hashstring was
getting malloc'ed in a loop but not freed.

So this one liner should mean the ketama example library gets cleaned up
properly.